all tournaments report and pack html report views in same module
