+++ flake8_rapport/index.html
@@ -9,7 +9,7 @@
       <div id="masthead" class="sev-4"></div>
       <div id="page">
           <h1>flake8 violations</h1>
-         <p id="versions">Generated on 2024-06-26 13:51
+         <p id="versions">Generated on 2024-06-26 14:12
             with Installed plugins: flake8-html: 0.4.3, mccabe: 0.7.0, pycodestyle: 2.12.0, pyflakes: 3.2.0
          </p>
          <ul id="index">
@@ -20,7 +20,7 @@
                   <span class="tick">&#x2713;</span>
                </span>
                <h2>All good!</h2>
-               <p>No flake8 errors found in 46 files scanned.</p>
+               <p>No flake8 errors found in 47 files scanned.</p>
             </div>
          </li>
